Bug fixing and updated the mask editor to use the newer editor component.

+++ backend/src/utils/expiration_manager.js
@@ -34,28 +34,26 @@ const expireNormalMap = async (id, force=false) => {
         const expiration_time = normalMap.updatedAt - new Date() + EXPIRE_DELAY * 60 * 1000
         if (force || expiration_time <= 0) {
             const images = normalMap.icon ? [...normalMap.layers, normalMap.icon] : normalMap.layers
-            console.log(images)
             for (var i=0; i < images.length; i++) {
                 const image = await Image.findById(images[i])
+                if (!image) { continue }
                 if (fs.existsSync(image.file)) {
                     fs.unlinkSync(image.file)
                     info(`Image ${image.file} has been deleted.`)
                 }
                 await Image.findByIdAndDelete(images[i])
             }
-            const creator = await User.findById(normalMap.creator.toString())
-            if (creator) {
-                const nm_index = creator.normalMaps.findIndex((nm) => nm.toString() === id.toString())
-                creator.normalMaps.splice(nm_index, 1)
-                await creator.save()
-            }
+            await User.updateOne(
+                { _id: normalMap.creator },
+                { $pull: { normalMaps: id } }
+            )
             await NormalMap.findByIdAndDelete(id)
             info('Normal map', id, 'deleted')
         } else {
             info('Normal map', id, 'will be deleted in', expiration_time/1000, 'seconds.')
             setTimeout(async () => await expireNormalMap(id), expiration_time)
         }
-    }
+    } 
 }
 
 const expireNormalMaps = async () => {
